+++ src/ngx_http_haskell_standalone.c
@@ -870,23 +870,16 @@ ngx_string(
 "ngxCyclePtr :: IO (AUX_NGX.Ptr ())\n"
 "ngxCyclePtr = AUX_NGX.readIORef aux_ngx_ngxCyclePtrStore\n\n"
 
-"aux_ngx_ngxCyclePtrStore :: AUX_NGX.IORef (AUX_NGX.Ptr ())\n"
-"aux_ngx_ngxCyclePtrStore =\n"
-"    AUX_NGX.unsafePerformIO $ AUX_NGX.newIORef AUX_NGX.nullPtr\n"
-"{-# NOINLINE aux_ngx_ngxCyclePtrStore #-}\n\n"
-
-"foreign export ccall ngxExportSetCyclePtr :: AUX_NGX.Ptr () -> IO ()\n"
-"ngxExportSetCyclePtr :: AUX_NGX.Ptr () -> IO ()\n"
-"ngxExportSetCyclePtr = AUX_NGX.writeIORef aux_ngx_ngxCyclePtrStore\n\n"
-
 "ngxUpstreamMainConfPtr :: IO (AUX_NGX.Ptr ())\n"
 "ngxUpstreamMainConfPtr = AUX_NGX.readIORef aux_ngx_ngxUpstreamMainConfPtrStore"
 "\n\n"
 
-"aux_ngx_ngxUpstreamMainConfPtrStore :: AUX_NGX.IORef (AUX_NGX.Ptr ())\n"
-"aux_ngx_ngxUpstreamMainConfPtrStore =\n"
-"    AUX_NGX.unsafePerformIO $ AUX_NGX.newIORef AUX_NGX.nullPtr\n"
-"{-# NOINLINE aux_ngx_ngxUpstreamMainConfPtrStore #-}\n\n"
+"ngxCachedTimePtr :: IO (AUX_NGX.Ptr ())\n"
+"ngxCachedTimePtr = AUX_NGX.readIORef aux_ngx_ngxCachedTimePtrStore\n\n"
+
+"foreign export ccall ngxExportSetCyclePtr :: AUX_NGX.Ptr () -> IO ()\n"
+"ngxExportSetCyclePtr :: AUX_NGX.Ptr () -> IO ()\n"
+"ngxExportSetCyclePtr = AUX_NGX.writeIORef aux_ngx_ngxCyclePtrStore\n\n"
 
 "foreign export ccall ngxExportSetUpstreamMainConfPtr :: AUX_NGX.Ptr () -> IO "
 "()\n"
@@ -894,5 +887,24 @@ ngx_string(
 "ngxExportSetUpstreamMainConfPtr =\n"
 "    AUX_NGX.writeIORef aux_ngx_ngxUpstreamMainConfPtrStore\n\n"
 
+"foreign export ccall ngxExportSetCachedTimePtr :: AUX_NGX.Ptr () -> IO ()\n"
+"ngxExportSetCachedTimePtr :: AUX_NGX.Ptr () -> IO ()\n"
+"ngxExportSetCachedTimePtr = AUX_NGX.writeIORef aux_ngx_ngxCachedTimePtrStore"
+"\n\n"
+
+"aux_ngx_ngxCyclePtrStore :: AUX_NGX.IORef (AUX_NGX.Ptr ())\n"
+"aux_ngx_ngxCyclePtrStore =\n"
+"    AUX_NGX.unsafePerformIO $ AUX_NGX.newIORef AUX_NGX.nullPtr\n"
+"{-# NOINLINE aux_ngx_ngxCyclePtrStore #-}\n\n"
+
+"aux_ngx_ngxUpstreamMainConfPtrStore :: AUX_NGX.IORef (AUX_NGX.Ptr ())\n"
+"aux_ngx_ngxUpstreamMainConfPtrStore =\n"
+"    AUX_NGX.unsafePerformIO $ AUX_NGX.newIORef AUX_NGX.nullPtr\n"
+"{-# NOINLINE aux_ngx_ngxUpstreamMainConfPtrStore #-}\n\n"
+
+"aux_ngx_ngxCachedTimePtrStore :: AUX_NGX.IORef (AUX_NGX.Ptr ())\n"
+"aux_ngx_ngxCachedTimePtrStore =\n"
+"    AUX_NGX.unsafePerformIO $ AUX_NGX.newIORef AUX_NGX.nullPtr\n"
+"{-# NOINLINE aux_ngx_ngxCachedTimePtrStore #-}\n\n"
 );